+++ src/dispatching/dialogue/mod.rs
@@ -33,10 +33,17 @@
 //! # #[cfg(feature = "macros")] {
 //! use std::convert::Infallible;
 //!
-//! use teloxide::{dispatching::dialogue::Transition, prelude::*, teloxide, RequestError};
+//! use teloxide::{
+//!     dispatching::dialogue::{InMemStorageError, Transition},
+//!     prelude::*,
+//!     teloxide, RequestError,
+//! };
 //!
+//! #[derive(Clone)]
 //! struct _1State;
+//! #[derive(Clone)]
 //! struct _2State;
+//! #[derive(Clone)]
 //! struct _3State;
 //!
 //! type Out = TransitionOut<D, RequestError>;
@@ -56,7 +63,7 @@
 //!     todo!()
 //! }
 //!
-//! #[derive(Transition)]
+//! #[derive(Clone, Transition)]
 //! enum D {
 //!     _1(_1State),
 //!     _2(_2State),
@@ -69,7 +76,7 @@
 //!     }
 //! }
 //!
-//! type In = DialogueWithCx<AutoSend<Bot>, Message, D, Infallible>;
+//! type In = DialogueWithCx<AutoSend<Bot>, Message, D, InMemStorageError>;
 //!
 //! #[tokio::main]
 //! async fn main() {
@@ -168,4 +175,4 @@ pub use storage::{RedisStorage, RedisStorageError};
 #[cfg(feature = "sqlite-storage")]
 pub use storage::{SqliteStorage, SqliteStorageError};
 
-pub use storage::{serializer, InMemStorage, Serializer, Storage, TraceStorage};
+pub use storage::{serializer, InMemStorage, InMemStorageError, Serializer, Storage, TraceStorage};